moves resources from `internal/lib/dynatrace` to `internal/sli` folders
