+++ internal/sli/dashboard.go
@@ -1,4 +1,4 @@
-package dynatrace
+package sli
 
 import (
 	"encoding/json"
