+++ internal/sli/dashboardlink.go
@@ -1,4 +1,4 @@
-package dynatrace
+package sli
 
 import (
 	"fmt"
