+++ internal/sli/dynatrace.go
@@ -1,4 +1,4 @@
-package dynatrace
+package sli
 
 import (
 	"crypto/tls"
