+++ internal/sli/management_zone_filter.go
@@ -1,4 +1,4 @@
-package dynatrace
+package sli
 
 import "fmt"
