+++ internal/sli/testfiles/test_get_metrics_hostcpuusage.json
