+++ internal/sli/testfiles/test_get_metrics_workingsetsize.json
